# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Estate360 API
  version: 1.0.0
  description: Real estate marketplace for Freetown, Sierra Leone.
paths:
...
  /api/v1/admin/listings/{id}/decision:
    post:
      operationId: admin_listings_decision_create
      description: POST /admin/listings/{id}/decision
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/v1/listings/{id}/submit:
    post:
      operationId: listings_submit_create
      description: POST /listings/{id}/submit
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - listings
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/v1/saved/:
    get:
      operationId: saved_retrieve
      description: GET /saved
      tags:
      - saved
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT